+++ extend/bajo/hook/waibu@on-request.js
@@ -1,8 +1,8 @@
 const onRequest = {
   level: 10,
   handler: async function (req, reply) {
-    const hostname = req.hostname.split(':')[0]
-    req.site = await this.getSite(hostname)
+    const { getHostname } = this.app.waibu
+    req.site = await this.getSite(getHostname(req))
   }
 }